Cleaned up cell states as read from the server a little
client.js had a pretty useless cell state enum, which didn't match
up to the server's... even though GameCell.changeState relied upon
the server's state names directly.

Replaced the enum in client.js used as GameCell.state values, with
the direct input values for changeState. Also added a server-to-client
lookup, which is purely for descriptive purposes, as the state name
strings are the same.