Deal with invalid "to clear" state in client
If a request is sent twice before the server responds, two turns
will be sent back: the real turn, and another with no cells cleared
but the same requested.

This screwed up highlighting of "to clear" cell going forwards through
turns, as it relied upon the turn afterwards setting the toClear cells
to either "mine" or "cleared".

toClear is no longer an entry in changeState, but simply an added/removed
CSS class, so other information about the cell is not lost. Then the toClear
class is just removed when navigating to a different turn.